Refactor ValidationRules folder structure
