+++ server/src/configTree.ts
@@ -1,26 +1,13 @@
-import { Diagnostic, DiagnosticSeverity } from "vscode-languageserver";
+import { Diagnostic } from "vscode-languageserver";
 import { ConfigTreeValidator } from "./configTreeValidator";
-import { uselessScope } from "./messageUtil";
-import { getRequirement } from "./relatedSettings";
-import { Requirement } from "./requirement";
+import { getRequirement, RequiredSettingsValidator } from "./requiredSettingsValidator";
 import { isNestedToPrevious, sectionDepthMap } from "./resources";
 import { Section } from "./section";
 import { Setting } from "./setting";
 import { TextRange } from "./textRange";
-import { createDiagnostic, getSetting } from "./util";
-
-export interface RelatedSettingsRule {
-    name: string;
-    rule: (section: Section, tree: ConfigTree) => Diagnostic | void;
-}
-
-export interface ValidationRule {
-    name: string;
-    rules: RelatedSettingsRule[];
-}
 
 /**
- * Checks related settings.
+ * Builds config tree structure
  */
 export class ConfigTree {
     public diagnostics: Diagnostic[];
@@ -29,6 +16,7 @@ export class ConfigTree {
     private previous: Section;
 
     private configTreeValidator: ConfigTreeValidator = new ConfigTreeValidator();
+    private requiredSettingsValidator: RequiredSettingsValidator = new RequiredSettingsValidator();
 
     public constructor(diagnostics: Diagnostic[]) {
         this.diagnostics = diagnostics;
@@ -112,7 +100,7 @@ export class ConfigTree {
                 /**
                  * Section contains dependent which can be useless or requires additional setting.
                  */
-                this.checkCurrentAndSetRequirementsForChildren(requirement, section, setting);
+                this.requiredSettingsValidator.checkCurrentAndSetRequirementsForChildren(requirement, section, setting);
             }
         }
     }
@@ -130,117 +118,16 @@ export class ConfigTree {
             let childAccumulator: Section[] = [];
             for (const parentSection of currentLevel) {
                 for (let [sectionToCheck, reqsForSection] of parentSection.requirementsForChildren) {
-                    this.checkAllChildren(sectionToCheck, reqsForSection, parentSection);
+                    this.requiredSettingsValidator.checkAllChildren(sectionToCheck, reqsForSection, parentSection);
                 }
                 childAccumulator.push(...parentSection.children);
             }
             currentLevel = childAccumulator;
         }
-
+        // Adds diagnostic about useless or missing required settings
+        this.diagnostics.push(...this.requiredSettingsValidator.diagnostics);
+        // Adds diagnostic of settings validation
         const diagnostic: Diagnostic[] =  this.configTreeValidator.validate(this);
         this.diagnostics.push(...diagnostic);
-    }
-
-    /**
-     * Checks each children of the section recursevly.
-     * @param targetSectionName name of the section, which need to be checked
-     * @param requirements array of requirements for the section, which need to be checked
-     * @param startSection root of subtree to search for targetSection
-     */
-    private checkAllChildren(targetSectionName: string, requirements: Requirement[], startSection: Section) {
-        if (startSection.children.length < 1) {
-            return;
-        }
-        for (const child of startSection.children) {
-            if (child.name === targetSectionName) {
-                this.checkSection(requirements, child);
-                continue;
-            }
-            this.checkAllChildren(targetSectionName, requirements, child);
-        }
-    }
-
-    /**
-     * Adds Diagnostic if section matches condifitons and doesn't contain required ("checked") setting.
-     */
-    private checkSection(requirements: Requirement[], section: Section) {
-        for (const req of requirements) {
-            if (section.sectionMatchConditions(req.conditions)) {
-                if (req.requiredIfConditions) {
-                    const required = req.requiredIfConditions;
-                    const checkedSetting = section.getSettingFromTree(required);
-                    const defaultValue = getSetting(required).defaultValue;
-                    if (checkedSetting == null && defaultValue == null) {
-                        this.diagnostics.push(createDiagnostic(section.range.range,
-                            `${required} is required if ${req.dependent} is specified`));
-                        return;
-                    }
-                } else {
-                    const anyRequiredIsSpecified = req.requiredAnyIfConditions.some(
-                        reqSetting => section.getSettingFromTree(reqSetting) != null);
-                    if (!anyRequiredIsSpecified) {
-                        this.diagnostics.push(createDiagnostic(section.range.range,
-                            `${req.dependent} has effect only with one of the following:
- * ${req.requiredAnyIfConditions.join("\n * ")}`));
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * If section doesn't match at least one condition, adds new Diagnostic about dependent.
-     */
-    private checkDependentUseless(section: Section, requirement: Requirement, dependent: Setting) {
-        const msg: string[] = requirement.conditions.map(condition => condition(section) as string).filter(m => m);
-        if (msg.length > 0) {
-            this.diagnostics.push(createDiagnostic(
-                dependent.textRange,
-                uselessScope(dependent.displayName, `${msg.join(", ")}`),
-                DiagnosticSeverity.Warning
-            ));
-        }
-    }
-
-    /**
-     * @param requirement Requirement for `dependent` setting.
-     * @param section Section, where `dependent` is declared.
-     * @param dependent Setting, which requires other settings or which must be checked for applicability.
-     */
-    private checkCurrentAndSetRequirementsForChildren(requirement: Requirement, section: Section, dependent: Setting) {
-        if (requirement.requiredIfConditions == null && requirement.requiredAnyIfConditions == null) {
-            this.checkDependentUseless(section, requirement, dependent);
-            return;
-        }
-        let requiredSetting;
-        if (requirement.requiredIfConditions) {
-            requiredSetting = getSetting(requirement.requiredIfConditions);
-        } else {
-            /**
-             * If requirement.requiredIfConditions == null, then requiredAnyIfConditions != null.
-             * It's supposed that all settings from `requiredAnyIfConditions` have the same sections,
-             * that's why only first section is used here.
-             */
-            requiredSetting = getSetting(requirement.requiredAnyIfConditions[0]);
-        }
-
-        const sectionNames = requiredSetting.section;
-        if (!sectionNames) {
-            return;
-        }
-        if (sectionNames.includes(section.name)) {
-            // check current
-            this.checkSection([requirement], section);
-            return;
-        }
-        const childSectionNames: string[] = typeof sectionNames === "string" ? [sectionNames] : sectionNames;
-        for (const secName of childSectionNames) {
-            const reqs = section.requirementsForChildren.get(secName);
-            if (reqs) {
-                reqs.push(requirement);
-            } else {
-                section.requirementsForChildren.set(secName, [requirement]);
-            }
-        }
     }
 }
