+++ server/src/validationRules/colorsThresholds.ts
@@ -1,9 +1,9 @@
 import { Diagnostic } from "vscode-languageserver";
-import { RelatedSettingsRule } from "../configTree";
 import { incorrectColors } from "../messageUtil";
 import { sectionMatchConditionRequired } from "../requirement";
 import { Section } from "../section";
 import { createDiagnostic } from "../util";
+import { RelatedSettingsRule } from "./interfaces";
 
 const rule: RelatedSettingsRule = {
     name: "Check colors match thresholds",
