+++ server/src/validationRules/index.ts
@@ -1,8 +1,7 @@
-import { ValidationRule } from "../configTree";
-
 import colorsThresholds from "./colorsThresholds";
 import forecastEndTime from "./forecastEndTime";
 import forecastSSA from "./forecastSSA";
+import { ValidationRule } from "./interfaces";
 import startEndTime from "./startEndTime";
 
 const rulesBySection: ValidationRule[] = [
